- robustness: improved docker builds

+++ openshift/build-docker-images.sh
@@ -1,11 +1,17 @@
 #!/bin/sh
 
-rm -rf _snafu_clean_
-git clone .. _snafu_clean_
-#cd _snafu_clean_
-cp Dockerfile* _snafu_clean_
-docker build -t snafu -f _snafu_clean_/Dockerfile _snafu_clean_
-docker build -t snafucomplete -f _snafu_clean_/Dockerfile.complete _snafu_clean_
+if [ "$1" = "build" ]
+then
+	rm -rf _snafu_clean_
+	git clone .. _snafu_clean_
+	#cd _snafu_clean_
+	cp Dockerfile* _snafu_clean_
+	docker build -t snafu -f _snafu_clean_/Dockerfile _snafu_clean_
+	docker tag snafu jszhaw/snafu
+	docker build -t snafucomplete -f _snafu_clean_/Dockerfile.complete _snafu_clean_
+else
+	echo "# Note: Not building. Run with 'build' option to build."
+fi
 
 if [ "$1" = "release" ]
 then
